Address android_lint.
