+++ packages/video_player/video_player_android/android/src/main/java/io/flutter/plugins/videoplayer/RemoteVideoAsset.java
@@ -1,0 +1,97 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+package io.flutter.plugins.videoplayer;
+
+import android.content.Context;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.annotation.OptIn;
+import androidx.annotation.VisibleForTesting;
+import androidx.media3.common.MediaItem;
+import androidx.media3.common.MimeTypes;
+import androidx.media3.common.util.UnstableApi;
+import androidx.media3.datasource.DataSource;
+import androidx.media3.datasource.DefaultDataSource;
+import androidx.media3.datasource.DefaultHttpDataSource;
+import androidx.media3.exoplayer.source.DefaultMediaSourceFactory;
+import androidx.media3.exoplayer.source.MediaSource;
+import java.util.Map;
+
+final class RemoteVideoAsset extends VideoAsset {
+  private static final String DEFAULT_USER_AGENT = "ExoPlayer";
+  private static final String HEADER_USER_AGENT = "User-Agent";
+
+  @NonNull private final StreamingFormat streamingFormat;
+  @NonNull private final Map<String, String> httpHeaders;
+
+  RemoteVideoAsset(
+      @Nullable String assetUrl,
+      @NonNull StreamingFormat streamingFormat,
+      @NonNull Map<String, String> httpHeaders) {
+    super(assetUrl);
+    this.streamingFormat = streamingFormat;
+    this.httpHeaders = httpHeaders;
+  }
+
+  @NonNull
+  @Override
+  MediaItem getMediaItem() {
+    MediaItem.Builder builder = new MediaItem.Builder().setUri(assetUrl);
+    String mimeType = null;
+    switch (streamingFormat) {
+      case Smooth:
+        mimeType = MimeTypes.APPLICATION_SS;
+        break;
+      case DynamicAdaptive:
+        mimeType = MimeTypes.APPLICATION_MPD;
+        break;
+      case HttpLive:
+        mimeType = MimeTypes.APPLICATION_M3U8;
+        break;
+    }
+    if (mimeType != null) {
+      builder.setMimeType(mimeType);
+    }
+    return builder.build();
+  }
+
+  @Override
+  MediaSource.Factory getMediaSourceFactory(Context context) {
+    return getMediaSourceFactory(context, new DefaultHttpDataSource.Factory());
+  }
+
+  /**
+   * Returns a configured media source factory, starting at the provided factory.
+   *
+   * <p>This method is provided for ease of testing without making real HTTP calls.
+   *
+   * @param context application context.
+   * @param initialFactory initial factory, to be configured.
+   * @return configured factory, or {@code null} if not needed for this asset type.
+   */
+  @VisibleForTesting
+  MediaSource.Factory getMediaSourceFactory(
+      Context context, DefaultHttpDataSource.Factory initialFactory) {
+    String userAgent = DEFAULT_USER_AGENT;
+    if (!httpHeaders.isEmpty() && httpHeaders.containsKey(HEADER_USER_AGENT)) {
+      userAgent = httpHeaders.get(HEADER_USER_AGENT);
+    }
+    unstableUpdateDataSourceFactory(initialFactory, httpHeaders, userAgent);
+    DataSource.Factory dataSoruceFactory = new DefaultDataSource.Factory(context, initialFactory);
+    return new DefaultMediaSourceFactory(context).setDataSourceFactory(dataSoruceFactory);
+  }
+
+  // TODO: Migrate to stable API, see https://github.com/flutter/flutter/issues/147039.
+  @OptIn(markerClass = UnstableApi.class)
+  private static void unstableUpdateDataSourceFactory(
+      @NonNull DefaultHttpDataSource.Factory factory,
+      @NonNull Map<String, String> httpHeaders,
+      @Nullable String userAgent) {
+    factory.setUserAgent(userAgent).setAllowCrossProtocolRedirects(true);
+    if (!httpHeaders.isEmpty()) {
+      factory.setDefaultRequestProperties(httpHeaders);
+    }
+  }
+}
